Rework stringio workaround
In JRuby 9K, the Java class was moved, and instead of writing too many
lines of Java-reflection code to keep the logic, the same logic was
implemented by dispatching through Ruby means.

I haven't verified whether the workaround is even necessary in JRuby 9K,
so it might turn out that this is the wrong thing to do.